Move s6 into rct2 header
